+++ src/main/java/woowacourse/shoppingcart/ui/ProductController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import woowacourse.shoppingcart.domain.Product;
 import woowacourse.shoppingcart.dto.Request;
-import woowacourse.shoppingcart.application.ProductService;
+import woowacourse.shoppingcart.service.ProductService;
 
 import java.net.URI;
 import java.util.List;
