+++ src/test/java/woowacourse/shoppingcart/ui/OrderControllerTest.java
@@ -12,7 +12,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import woowacourse.shoppingcart.domain.OrderDetail;
 import woowacourse.shoppingcart.dto.OrderRequest;
 import woowacourse.shoppingcart.domain.Orders;
-import woowacourse.shoppingcart.application.OrderService;
+import woowacourse.shoppingcart.service.OrderService;
 
 import java.util.Arrays;
 import java.util.Collections;
